% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\name{.Get_percent_predicted}
\alias{.Get_percent_predicted}
\title{Get \% of expected value}
\usage{
.Get_percent_predicted(
  metric,
  sex,
  level,
  age,
  measurement,
  verbose = FALSE,
  digits = 0
)
}
\arguments{
\item{metric}{character, body composition metric}

\item{sex}{character, "Female" or "Male"}

\item{level}{character, used vertebral level}

\item{age}{integer, age}

\item{measurement}{numeric, raw value of measurement}

\item{verbose}{logical, should messages be discplayed}

\item{digits}{integer, digits to round return value}
}
\value{
numeric, corresponding percentile
}
\description{
For a given constellation of metric, sex, vertebral level, and age returns the ratio of actual and expected value (percentile 50) in percent
}
\author{
J. Peter Marquardt
}
