% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\name{.Get_reference_model}
\alias{.Get_reference_model}
\title{Get reference models}
\usage{
.Get_reference_model(fitted_metric, sex, level)
}
\arguments{
\item{fitted_metric}{string, name of fitted metric (abbreviation, includes a lowercase b prefix for metrics with box-cox-transformation)}

\item{sex}{character, "Male" or "Female"}

\item{level}{string, vertebral level, T5, T8, T10, or L3}
}
\value{
LMSP model
}
\description{
For a given combination of metric, sex, and level return the appropriate LMSP model.
Imports the non-CRAN packages musclerefdata and adiposerefdata
}
\author{
J.Peter Marquardt
}
