% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\name{.Get_reference_percentile}
\alias{.Get_reference_percentile}
\title{Get percentile for a given combination of model and measurement}
\usage{
.Get_reference_percentile(
  metric,
  sex,
  level,
  age,
  measurement,
  verbose = FALSE,
  digits = 0
)
}
\arguments{
\item{metric}{character, body composition metric}

\item{sex}{character, "Female" or "Male"}

\item{level}{character, used vertebral level}

\item{age}{integer, age}

\item{measurement}{numeric, raw value of measurement}

\item{verbose}{logical, should messages be displayed}

\item{digits}{integer, digits to round percentile to}
}
\value{
numeric, corresponding percentile
}
\description{
For a given constellation of metric, sex, vertebral level, and age returns the corresponding percentile for a given measurement
}
\author{
J. Peter Marquardt
}
