% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\name{reference_z_scores}
\alias{reference_z_scores}
\title{Get z-scores for a given combination of model and measurement}
\usage{
reference_z_scores(
  metric,
  sex,
  level,
  age,
  measurement,
  verbose = FALSE,
  digits = 2
)
}
\arguments{
\item{metric}{character (vector), body composition metric}

\item{sex}{character (vector), "Female" or "Male"}

\item{level}{character (vector), used vertebral level}

\item{age}{integer (vector), age}

\item{measurement}{numeric (vector), raw value of measurement}

\item{verbose}{logical, should messages be displayed}

\item{digits}{integer, digits to round percentile to}
}
\value{
numeric, corresponding z-score
}
\description{
For a given constellation of metric, sex, vertebral level, and age returns the corresponding z-scores for a given set of measurements
}
\seealso{
[bodycomp_reference()]
}
\author{
J. Peter Marquardt
}
